+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changelog
 3.0.5 (unreleased)
 ------------------
 
+- User that manage a rule can now allow actions executed by this rule
+  to recursively trigger other rules.
+  For example, if you have a rule that automatically publish added content,
+  and an other rule that sends an email when a content is published,
+  if autopublish rule is marked as 'recursive', then send mail rule will be triggered.
+  [thomasdesvenain]
+
 - Mail action: if string interpolation of recipient gives several times
   the same email address,
   the email is sent only once to this recipient.